+++ CoinCap/Entities/Assets/HistoryById.cs
@@ -5,7 +5,7 @@ namespace CoinCap.Entities.Assets
     public class HistoryById
     {
         [JsonPropertyName("priceUsd")]
-        public double? PriceUsd { get; set; }
+        public decimal? PriceUsd { get; set; }
 
         [JsonPropertyName("time")]
         public long? Time { get; set; }
