+++ CoinCap/Entities/Candles/Candle.cs
@@ -5,19 +5,19 @@ namespace CoinCap.Entities.Candles
     public class Candle
     {
         [JsonPropertyName("open")]
-        public double? Open { get; set; }
+        public decimal? Open { get; set; }
 
         [JsonPropertyName("high")]
-        public double? High { get; set; }
+        public decimal? High { get; set; }
 
         [JsonPropertyName("low")]
-        public double? Low { get; set; }
+        public decimal? Low { get; set; }
 
         [JsonPropertyName("close")]
-        public double? Close { get; set; }
+        public decimal? Close { get; set; }
 
         [JsonPropertyName("volume")]
-        public double? Volume { get; set; }
+        public decimal? Volume { get; set; }
 
         [JsonPropertyName("period")]
         public long? Period { get; set; }
